requirements.txt
python-dotenv
selenium
schedule
cryptography
